Squashed commit of the following:
commit 06d3bdcc55a96e776c838df76e92cc87341d060e
Author: Wolfleader <25214835+Wolfleader101@users.noreply.github.com>
Date:   Sat Apr 15 15:50:53 2023 +0800

    Lua entity access tea 54 (#24)

    * adding in time to lua

    * fixed some boo boos

    * Working on adding in functionality to lua

    * system update

    * fixed issues

    * removed log

    * fixed some stuff

    * change VAO from float to Vertex

    * fixes

    * trying

    * works

    * maybe drawing works?

    * use assetloader like this, we can fix after

    * example

    * Added in camera

    * Working on input in lua

    * camera controller in lua

    * pitch movement

    * working on getting camera moving

    * might have something working DO NOT PULL!

    * fixed getting delta pos

    * step 1

    * fixed script issue

    * Update basic.vert

    * made it local

    * lua changes

    * changes

    * updates

    * Update GLRenderer.cpp

    * model material + texture support

    * obj fix

    * adding in scaling and transform

    * cube fix

    * fixing issue

    * assets

    * add phong

    * trying to fix error

    * fixed issues

    * more textures + emssision

    * add point light

    * spot light

    * light cookies

    * merging with lua branch

    * instancing

    * trying to fix basic demo

    * ch.ch.ch.ch ch changes

    * Update GLInstance.cpp

    * whoopsies

    ---------

+++ Core/src/Controller/Time.hpp
@@ -9,29 +9,28 @@ namespace Vakol {
          * @brief time data used everywhere
          */
         struct Time {
-
-            Time() {};
+            Time(){};
             /**
              * @brief current time
              */
-            double curTime = 0.0f;
+            float curTime = 0.0f;
 
             /**
              * @brief delta time
              */
-            double deltaTime = 0.0f;
+            float deltaTime = 0.0f;
 
             /**
              * @brief the previous frames time
              *
              */
-            double prevTime = 0.0f;
+            float prevTime = 0.0f;
 
             /**
              * @brief used to update the time struct
              */
             void Update() {
-                curTime = glfwGetTime();
+                curTime = static_cast<float>(glfwGetTime());
                 if (prevTime == 0.0f)
                     deltaTime = 0.0f;
                 else
@@ -39,5 +38,5 @@ namespace Vakol {
                 prevTime = curTime;
             }
         };
-    }
-}  // namespace Vakol::Controller
+    }  // namespace Controller
+}  // namespace Vakol
